+++ src/main/java/bowling/domain/player/Players.java
@@ -1,5 +1,6 @@
-package bowling.domain;
+package bowling.domain.player;
 
+import bowling.domain.player.Player;
 import java.util.Collections;
 import java.util.List;
