+++ src/main/java/bowling/view/InputView.java
@@ -1,6 +1,6 @@
 package bowling.view;
 
-import bowling.domain.Name;
+import bowling.domain.player.Name;
 import java.util.Scanner;
 
 public class InputView {
