+++ src/test/java/bowling/domain/PlayerTest.java
@@ -2,6 +2,7 @@ package bowling.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import bowling.domain.player.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
